add all home page tests.
