+++ frontend/ai.client/e2e/auth/navigation.spec.ts
