+++ frontend/ai.client/e2e/home-page/file-upload-ui.user.spec.ts
