+++ frontend/ai.client/e2e/home-page/model-selector.user.spec.ts
