+++ frontend/ai.client/e2e/home-page/settings-panel.user.spec.ts
