+++ frontend/ai.client/src/app/components/sidenav/sidenav.html
@@ -9,7 +9,7 @@
             [appTooltip]="'Close sidebar'"
             appTooltipPosition="right"
             class="hidden lg:flex items-center justify-center size-8 rounded-md text-gray-500 hover:bg-gray-200 hover:text-gray-700 dark:text-gray-400 dark:hover:bg-white/10 dark:hover:text-white"
-            [attr.aria-label]="isCollapsed() ? 'Expand sidebar' : 'Collapse sidebar'">
+            aria-label="Collapse sidebar">
             <!-- Sidebar panel icon -->
             <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.5" class="size-6" aria-hidden="true">
                 <rect x="3" y="4" width="18" height="16" rx="2" ry="2" />
